+++ odoo/local-src/db2_import/models/db2_importer.py
@@ -322,6 +322,32 @@ class DB2MapperSaleOrder(object):
 
         create_date = convert_date('eccc', row)
         customer = rec.env.ref(convert_customer(int(row['ecccli'])))
+        pricelist = customer.property_product_pricelist
+        pay_term = customer.property_payment_term_id
+        addr = customer.address_get(['delivery', 'invoice'])
+
+        delivery = rec.env['res.partner'].browse(addr['delivery'])
+        # take fiscal posistion to work with a record
+        # partner manually set fiscal position always win
+        fpos = (delivery.property_account_position_id or
+                customer.property_account_position_id)
+        if not fpos:
+            fp_obj = rec.env['account.fiscal.position']
+            # First search only matching VAT positions
+            vat_required = bool(customer.vat)
+            fpos = fp_obj._get_fpos_by_region(
+                delivery.country_id.id,
+                delivery.state_id.id,
+                delivery.zip,
+                vat_required)
+
+            # Then if VAT required found no match, try positions that do not
+            # require it
+            if not fpos and vat_required:
+                fpos = fp_obj._get_fpos_by_region(
+                    delivery.country_id.id,
+                    delivery.state_id.id, delivery.zip, False)
+
         promo_sale = customer.supplier_promotion_sale_allowed
 
         user_xmlid = convert_user(row['eccrep'])
@@ -336,6 +362,11 @@ class DB2MapperSaleOrder(object):
             'confirmation_date': convert_date('eccd', row),
             'write_date': convert_date('eccm', row) or create_date,
             'partner_id': customer.id,
+            'pricelist_id': pricelist.id,
+            'payment_term_id': pay_term.id,
+            'partner_invoice_id': addr['invoice'],
+            'partner_shipping_id': delivery.id,
+            'fiscal_position_id': fpos.id,
             'supplier_promotion_allowed': promo_sale,
         }
 
@@ -385,6 +416,13 @@ class DB2MapperSaleOrder(object):
             if product_xmlid == '__setup__.product_other':
                 name = "Divers"
             product = rec.env.ref(product_xmlid)
+            # While odoo could do it for us on create
+            # in _prepare_add_missing_fields
+            # Do it ourselves to avoid call to onchange
+            taxes = product.taxes_id.filtered(
+                lambda r: r.company_id == rec.env.user.company_id)
+            taxes = fpos.map_tax(
+                taxes, product, new.partner_shipping_id) if fpos else taxes
             create_date = convert_date('dccc', line)
             values = {
                 'order_id': new.id,
@@ -395,6 +433,7 @@ class DB2MapperSaleOrder(object):
                 'product_uom': rec.env.ref('product.product_uom_unit').id,
                 'qty_delivered': line['dccqul'],
                 'price_unit': line['dccpvd'],
+                'tax_id': [(4, tax.id) for tax in taxes],
                 'discount': line['dccrem'],
                 'create_date': create_date,
                 'write_date': convert_date('dccm', line) or create_date,